+++ mock/modules/authorization/index.js
@@ -1,8 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const UserSerializer = require("../serializers/UserSerializer");
+const UserSerializer = require("../../serializers/UserSerializer");
 
 router.route("/login").post((req, res) => {
+  //登录
   const token = "Bearer EGbbP8ZWU1u-7dogAj97N5gemefVyVpR_50eErvfssA";
   const setSendValue = role => {
     res.send({
@@ -33,6 +34,7 @@ router.route("/login").post((req, res) => {
 });
 
 router.route("/register").post((req, res) => {
+  //注册
   res.send({});
 });
