+++ mock/modules/user/index.js
@@ -1,9 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const UserSerializer = require("../serializers/UserSerializer");
-const TagSerializer = require("../serializers/TagSerializer");
+const UserSerializer = require("../../serializers/UserSerializer");
+const TagSerializer = require("../../serializers/TagSerializer");
 
 router.route("/:userId").get((req, res) => {
+  //获得用户信息
   res.send({
     data: UserSerializer(),
     abilities: {
@@ -13,6 +14,7 @@ router.route("/:userId").get((req, res) => {
 });
 
 router.route("/:userId").post((req, res) => {
+  //修改用户信息
   res.send({
     data: UserSerializer(),
     abilities: {
@@ -22,6 +24,7 @@ router.route("/:userId").post((req, res) => {
 });
 
 router.route("/:userId/password").post((req, res) => {
+  //修改密码
   res.send({
     data: UserSerializer(),
     abilities: {
@@ -31,6 +34,7 @@ router.route("/:userId/password").post((req, res) => {
 });
 
 router.route("/:userId/tag").post((req, res) => {
+  //添加标签
   res.send({
     data: TagSerializer()
   });
